refactor(//core/runtime): Renaming execution -> runtime
